backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
contourpy==1.3.3
cycler==0.12.1
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
Faker==40.1.2
fonttools==4.61.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==25.0
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
